requirements.txt
# Web Framework and Extensions
Flask==2.2.2
Werkzeug==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-WTF==1.2.1

# Database
SQLAlchemy==1.4.27
psycopg2-binary
alembic==1.13.1

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
argon2-cffi==23.1.0

# HTTP & Networking
requests==2.31.0
urllib3==2.1.0
httpx==0.26.0

# Task Queue & Cache
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utils
python-dotenv==1.0.0
PyYAML==6.0.1
python-dateutil==2.8.2

# Development Tools
black==23.12.1
flake8==7.0.0
